+++ src/test/java/BaseTest.java
@@ -1,8 +1,40 @@
 import io.github.bonigarcia.wdm.WebDriverManager;
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.chrome.ChromeOptions;
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.BeforeSuite;
+import org.openqa.selenium.WebDriver;
+
+import java.time.Duration;
+import java.util.UUID;
 
 public class BaseTest {
+
+    public static WebDriver driver = null;
+    public static String url = "https://qa.koel.app/";
+
+    @BeforeMethod
+    public void launchBrowser() {
+        //      Added ChromeOptions argument below to fix websocket error
+        ChromeOptions options = new ChromeOptions();
+        options.addArguments("--remote-allow-origins=*");
+
+        driver = new ChromeDriver(options);
+        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
+    }
+
+    @AfterMethod
+    public void closeBrowser() {
+        driver.quit();
+    }
+
+    protected static void openLoginUrl() {
+        String url = "https://qa.koel.app/";
+        driver.get(url);
+    }
 
     @BeforeSuite
     static void setupClass() {
