+++ src/test/java/LoginTests.java
@@ -10,15 +10,8 @@ public class LoginTests extends BaseTest {
     @Test
     public void LoginEmptyEmailPasswordTest() {
 
-//      Added ChromeOptions argument below to fix websocket error
-        ChromeOptions options = new ChromeOptions();
-        options.addArguments("--remote-allow-origins=*");
+        openLoginUrl();
 
-        WebDriver driver = new ChromeDriver(options);
-        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
-
-        String url = "https://qa.koel.app/";
-        driver.get(url);
         Assert.assertEquals(driver.getCurrentUrl(), url);
         driver.quit();
     }
