+++ src/test/resources/TestNG.xml
@@ -6,7 +6,7 @@
     <test name="Example Test" preserve-order="false">
         <classes>
             <class name="LoginTests"/>
-            <class name="Homework16"/>
+            <class name="RegistrationNavigation"/>
             <class name="BaseTest"/>
         </classes>
     </test>
